backend/requirements.txt
Flask
python-magic-bin
pypdf2
pydantic
langchain-deepseek
